now admin can see orders with their prescription

+++ lib/Controller/Customer Orders Screen Controller/customer_orders_screen_controller.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:medicart/Controller/Customer%20Orders%20Screen%20Controller/customer_orders_screen_state.dart';
@@ -23,10 +25,36 @@ class CustomerOrdersScreenController
         .collection('orders')
         .doc(userId)
         .collection('products');
-state = state.copywith(isloading: true);
-    //delete that product
-    await ordersRef.doc(orderedItemId).delete();
-    state = state.copywith(isloading: false);
+    final prescriptionRef = FirebaseFirestore.instance
+        .collection('prescriptions')
+        .doc(userId)
+        .collection('prescription');
+
+    state = state.copywith(isloading: true);
+    try {
+      //get unique code of ordereditem
+      final ordersSnapshot = await ordersRef.doc(orderedItemId).get();
+      if (ordersSnapshot.exists) {
+        final uniqueCode = ordersSnapshot.data()?['code'];
+
+        //delete the ordered item
+        await ordersRef.doc(orderedItemId).delete();
+
+        //delete prescription with same unique code
+        if (uniqueCode != null) {
+          final prescriptionSnapshot =
+              await prescriptionRef.where('code', isEqualTo: uniqueCode).get();
+
+          for (var doc in prescriptionSnapshot.docs) {
+            await prescriptionRef.doc(doc.id).delete();
+          }
+        }
+      }
+      state = state.copywith(isloading: false);
+    } catch (e) {
+      state = state.copywith(isloading: false);
+
+      log("Error deleting item or prescription: $e");
+    }
   }
-  
 }
